Fix unreachable code in Elastic plugin when processing like constraints. The intention of the unreachable code is to transform constraints of the form "like abc%" into ElasticSearch prefix queries instead of using the more expansive wildcard queries. This commit enables this optimization.
Move a couple of Lucene tests for the like constraint to oak-search module, so it is run in both Elastic and Lucene. One of these tests is ignored for Elastic because it is failing (see OAK-9885).
