+++ oak-lucene/src/test/java/org/apache/jackrabbit/oak/plugins/index/lucene/LucenePropertyIndexTest.java
@@ -1193,76 +1193,6 @@ public class LucenePropertyIndexTest extends AbstractQueryTest {
     }
 
     @Test
-    public void likeQueriesWithString() throws Exception {
-        Tree idx = createIndex("test1", of("propa", "propb"));
-        idx.addChild(PROP_NODE).addChild("propa");
-        root.commit();
-
-        Tree test = root.getTree("/").addChild("test");
-        test.addChild("a").setProperty("propa", "humpty");
-        test.addChild("b").setProperty("propa", "dumpty");
-        test.addChild("c").setProperty("propa", "humpy");
-        root.commit();
-
-        assertQuery("select [jcr:path] from [nt:base] where propa like 'hum%'",
-                asList("/test/a", "/test/c"));
-        assertQuery("select [jcr:path] from [nt:base] where propa like '%ty'",
-                asList("/test/a", "/test/b"));
-        assertQuery("select [jcr:path] from [nt:base] where propa like '%ump%'",
-                asList("/test/a", "/test/b", "/test/c"));
-    }
-
-    @Test
-    public void likeQueriesWithEscapedChars() throws Exception {
-        Tree idx = createIndex("test1", of("propa", "propb"));
-        idx.addChild(PROP_NODE).addChild("propa");
-        root.commit();
-
-        Tree test = root.getTree("/").addChild("test");
-        test.addChild("a").setProperty("propa", "foo%");
-        test.addChild("b").setProperty("propa", "%bar");
-        test.addChild("c").setProperty("propa", "foo%bar");
-        test.addChild("d").setProperty("propa", "foo_");
-        test.addChild("e").setProperty("propa", "_foo");
-        test.addChild("f").setProperty("propa", "foo_bar");
-        test.addChild("g").setProperty("propa", "foo%_bar");
-        test.addChild("h").setProperty("propa", "foo\\bar");
-        test.addChild("i").setProperty("propa", "foo\\\\%bar");
-        root.commit();
-
-        assertQuery("select [jcr:path] from [nt:base] where propa like 'foo%'",
-                asList("/test/a", "/test/c", "/test/d", "/test/f", "/test/g", "/test/h", "/test/i"));
-        assertQuery("select [jcr:path] from [nt:base] where propa like '%oo%'",
-                asList("/test/a", "/test/c", "/test/d", "/test/e", "/test/f", "/test/g", "/test/h", "/test/i"));
-        assertQuery("select [jcr:path] from [nt:base] where propa like 'foo\\%'",
-                asList("/test/a"));
-        assertQuery("select [jcr:path] from [nt:base] where propa like '%oo\\%'",
-                asList("/test/a"));
-        assertQuery("select [jcr:path] from [nt:base] where propa like '%oo\\%%'",
-                asList("/test/a", "/test/c", "/test/g"));
-        assertQuery("select [jcr:path] from [nt:base] where propa like '\\%b%'",
-                asList("/test/b"));
-        assertQuery("select [jcr:path] from [nt:base] where propa like 'foo_'",
-                asList("/test/a", "/test/d"));
-        assertQuery("select [jcr:path] from [nt:base] where propa like '_oo_'",
-                asList("/test/a", "/test/d"));
-        assertQuery("select [jcr:path] from [nt:base] where propa like 'foo\\_'",
-                asList("/test/d"));
-        assertQuery("select [jcr:path] from [nt:base] where propa like '%oo\\_'",
-                asList("/test/d"));
-        assertQuery("select [jcr:path] from [nt:base] where propa like '%oo\\_%'",
-                asList("/test/d", "/test/f"));
-        assertQuery("select [jcr:path] from [nt:base] where propa like '%oo\\%\\_%'",
-                asList("/test/g"));
-        assertQuery("select [jcr:path] from [nt:base] where propa like 'foo\\\\bar'",
-                asList("/test/h"));
-        assertQuery("select [jcr:path] from [nt:base] where propa like '%\\\\%'",
-                asList("/test/h", "/test/i"));
-        assertQuery("select [jcr:path] from [nt:base] where propa like '%\\\\\\%%'",
-                asList("/test/i"));
-    }
-
-    @Test
     public void nativeQueries() throws Exception {
         Tree idx = createIndex("test1", of("propa", "propb"));
         idx.addChild(PROP_NODE).addChild("propa");
